+++ BitigDataTests/XmlDirectionRepoTests.cs
@@ -5,6 +5,7 @@ using System.IO;
 using Bitig.Data.Storage;
 using Bitig.Logic.Model;
 using Bitig.Logic.Repository;
+using System.Collections.Generic;
 
 namespace BitigDataTests
 {
@@ -15,16 +16,19 @@ namespace BitigDataTests
         {
         }
 
-        private readonly string testFilePath = @"..\..\TestData\Directions.xml";
-        private readonly string alifbaPath = @"..\..\TestData\Alphabets.xml";
-        private readonly string preparedFile = @"..\..\TestData\Prepared\Direction777.xml";
+        private const string dataFolder = @"..\..\..\BitigDataTests\TestData\";
+        private readonly string alifbaPath = dataFolder + "Alphabets.xml";
+        private readonly string preparedFile = dataFolder + @"Prepared\Direction777.xml";
+        private readonly string testFilePath = dataFolder + "Directions.xml";
 
         [TestInitialize]
         [TestCleanup]
         public void DeleteTestFile()
         {
-            File.Delete(testFilePath);
-            File.Delete(alifbaPath);
+            if (File.Exists(testFilePath))
+                File.Delete(testFilePath);
+            if (File.Exists(alifbaPath))
+                File.Delete(alifbaPath);
         }
 
         [TestMethod]
@@ -41,7 +45,8 @@ namespace BitigDataTests
         public void GetList_CreateDefault()
         {
             var _alifbaRepo = new XmlAlifbaRepository(alifbaPath);
-            var _testRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _testRepo = new BitigDirectionRepository(
+                new XmlDirectionRepository(testFilePath, _alifbaRepo));
             var _builtInList = DefaultConfiguration.BuiltInDirections;
             Assert.AreNotEqual(0, _builtInList.Count);
             var _defaultList = _testRepo.GetList();
@@ -55,7 +60,7 @@ namespace BitigDataTests
         public void Get()
         {
             File.Copy(preparedFile, testFilePath);
-            var _alifbaRepo = new XmlAlifbaRepository(alifbaPath);
+            var _alifbaRepo = new BitigAlifbaRepository(new XmlAlifbaRepository(alifbaPath));
             var _testRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
             var _direction777 = _testRepo.Get(777);
             Assert.IsNotNull(_direction777);
@@ -73,20 +78,22 @@ namespace BitigDataTests
         public void Get_CreateDefault()
         {
             var _alifbaRepo = new XmlAlifbaRepository(alifbaPath);
-            var _testRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _xmlRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _testRepo = new BitigDirectionRepository(_xmlRepo);
             _testRepo.Get(0);
 
-            var _count = _testRepo.GetListNoCreateDefaults().Count;
+            var _count = _xmlRepo.GetList().Count;
             Assert.AreEqual(DefaultConfiguration.BuiltInDirections.Count, _count);
         }
 
         [TestMethod]
         public void Insert()
         {
-            var _alifbaRepo = new XmlAlifbaRepository(alifbaPath);
-            var _testRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
-            var _source = new Alifba(0, "alifba1");
-            var _target = new Alifba(1, "alifba2");
+            var _alifbaRepo = new BitigAlifbaRepository(new XmlAlifbaRepository(alifbaPath));
+            var _testRepo = new BitigDirectionRepository(
+                new XmlDirectionRepository(testFilePath, _alifbaRepo));
+            var _source = new Alifba(1, "alifba1");
+            var _target = new Alifba(0, "alifba2");
             var _assemblyName = "TestAssembly" + Guid.NewGuid();
             var _direction = new Direction(-1, _source, _target, _assemblyName, "TestType");
             _testRepo.Insert(_direction);
@@ -108,14 +115,16 @@ namespace BitigDataTests
         [TestMethod]
         public void Insert_AssignID()
         {
-            var _alifbaRepo = new XmlAlifbaRepository(alifbaPath);
-            var _testRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
-            var _source = new Alifba(0, "alifba1");
-            var _target = new Alifba(1, "alifba2");
+            var _alifbaRepo = new BitigAlifbaRepository(new XmlAlifbaRepository(alifbaPath));
+            var _testRepo = new BitigDirectionRepository(
+                new XmlDirectionRepository(testFilePath, _alifbaRepo));
+            var _source1 = new Alifba(1, "alifba1");
+            var _target = new Alifba(0, "alifba2");
+            var _source2 = new Alifba(2, "alifba3");
             const int _id = -1;
-            var _dir1 = new Direction(_id, _source, _target);
+            var _dir1 = new Direction(_id, _source1, _target);
             _testRepo.Insert(_dir1);
-            var _dir2 = new Direction(_id, _target, _source);
+            var _dir2 = new Direction(_id, _source2, _target);
             _testRepo.Insert(_dir2);
             Assert.AreNotEqual(-1, _dir1.ID);
             Assert.AreNotEqual(-1, _dir2.ID);
@@ -131,25 +140,48 @@ namespace BitigDataTests
         [TestMethod]
         public void Insert_CreateDefault()
         {
-            var _alifbaRepo = new XmlAlifbaRepository(alifbaPath);
-            var _testRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
-            var _source = new Alifba(0, "alifba1");
-            var _target = new Alifba(1, "alifba2");
+            var _alifbaRepo = new BitigAlifbaRepository(new XmlAlifbaRepository(alifbaPath));
+            var _xmlRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _testRepo = new BitigDirectionRepository(_xmlRepo);
+            var _source = new Alifba(1, "alifba1");
+            var _target = new Alifba(0, "alifba2");
             var _assemblyName = "TestAssembly" + Guid.NewGuid();
             var _direction = new Direction(-1, _source, _target, _assemblyName, "TestType");
             _testRepo.Insert(_direction);
             _testRepo.SaveChanges();
 
-            var _count = _testRepo.GetListNoCreateDefaults().Count;
+            var _count = _xmlRepo.GetList().Count;
             Assert.AreEqual(DefaultConfiguration.BuiltInDirections.Count + 1, _count);
+        }
+
+        [TestMethod]
+        public void Insert_SameAlifbaIDs()
+        {
+            var _alifbaRepo = new BitigAlifbaRepository(new XmlAlifbaRepository(alifbaPath));
+            var _testRepo = new BitigDirectionRepository(
+                new XmlDirectionRepository(testFilePath, _alifbaRepo));
+            var _source = new Alifba(0, "alifba1");
+            var _target = new Alifba(1, "alifba2");
+            var _assemblyName = "TestAssembly" + Guid.NewGuid();
+            var _direction = new Direction(-1, _source, _target, _assemblyName, "TestType");
+            try
+            {
+                _testRepo.Insert(_direction);
+                Assert.Fail("Shouldn't have inserted");
+            }
+            catch (Exception ex)
+            {
+                Assert.AreEqual("Direction with same source and target exists.", ex.Message);
+            }            
         }
 
         [TestMethod]
         public void Update()
         {
             File.Copy(preparedFile, testFilePath);
-            var _alifbaRepo = new XmlAlifbaRepository(alifbaPath);
-            var _testRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _alifbaRepo = new BitigAlifbaRepository(new XmlAlifbaRepository(alifbaPath));
+            var _testRepo = new BitigDirectionRepository(
+                new XmlDirectionRepository(testFilePath, _alifbaRepo));
             var _direction = _testRepo.Get(777);
             Assert.AreEqual(null, _direction.BuiltIn);
             Assert.AreEqual(0, _direction.Source.ID);
@@ -181,16 +213,17 @@ namespace BitigDataTests
         [TestMethod]
         public void Update_CreateDefault()
         {
-            var _alifbaRepo = new XmlAlifbaRepository(alifbaPath);
-            var _testRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _alifbaRepo = new BitigAlifbaRepository(new XmlAlifbaRepository(alifbaPath));
+            var _xmlRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _testRepo = new BitigDirectionRepository(_xmlRepo);
             var _assembly = "TestAssembly " + Guid.NewGuid();
             var _source = _alifbaRepo.Get(2);
             var _target = _alifbaRepo.Get(3);
             var _direction = new Direction(0, _source, _target, _assembly);
             _testRepo.Update(_direction);
             _testRepo.SaveChanges();
-            var _list = _testRepo.GetListNoCreateDefaults();
-            Assert.AreEqual(DefaultConfiguration.BuiltInAlifbaList.Count, _list.Count);
+            var _list = _xmlRepo.GetList();
+            Assert.AreEqual(DefaultConfiguration.BuiltInDirections.Count, _list.Count);
         }
 
         [TestMethod]
@@ -213,14 +246,13 @@ namespace BitigDataTests
         public void Delete_CreateDefault()
         {
             var _alifbaRepo = new XmlAlifbaRepository(alifbaPath);
-            var _testRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
-            //var _source = _alifbaRepo.Get(2);
-            //var _target = _alifbaRepo.Get(3);
+            var _xmlRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _testRepo = new BitigDirectionRepository(_xmlRepo);
             var _direction = new Direction(0, null, null); //repo: Repo.FirstID? (together with Repo.DefaultID)
             _testRepo.Delete(_direction);
             _testRepo.SaveChanges();
-            var _list = _testRepo.GetListNoCreateDefaults();
-            Assert.AreEqual(DefaultConfiguration.BuiltInAlifbaList.Count - 1, _list.Count);
+            var _list = _xmlRepo.GetList();
+            Assert.AreEqual(DefaultConfiguration.BuiltInDirections.Count - 1, _list.Count);
         }
 
         [TestMethod]
@@ -230,6 +262,30 @@ namespace BitigDataTests
             var _IDs = new int[] { 0, 1, 3, 4, 6 };
             var _newID = _testRepo.GenerateID(_IDs);
             Assert.IsFalse(_IDs.Any(_id => _id == _newID));
+        }
+
+        [TestMethod]
+        public void GetByAlifbaIDs()
+        {
+            var _alifbaRepo = new BitigAlifbaRepository(new XmlAlifbaRepository(alifbaPath));
+            var _xmlRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _testRepo = new BitigDirectionRepository(_xmlRepo);
+            var _cyrZam = _testRepo.GetByAlifbaIDs(0, 2);
+            Assert.IsNotNull(_cyrZam);
+            Assert.AreEqual(0, _cyrZam.Source.ID);
+            Assert.AreEqual(2, _cyrZam.Target.ID);
+        }
+
+        [TestMethod]
+        public void GetTargets()
+        {
+            var _alifbaRepo = new BitigAlifbaRepository(new XmlAlifbaRepository(alifbaPath));
+            var _xmlRepo = new XmlDirectionRepository(testFilePath, _alifbaRepo);
+            var _testRepo = new BitigDirectionRepository(_xmlRepo);
+            var _targets = _testRepo.GetTargets(0);
+            Assert.AreEqual(3, _targets.Count);
+            var _expectedIDs = new List<int> { 1, 2, 3 };
+            Assert.IsTrue(_expectedIDs.All(_id => _targets.Exists(_t => _t.ID == _id)));
         }
     }
 }
